fix: Token and GasPrice data is stored big-endian encoded in FeeQuoter
